Move data hooks to protocol
