+++ apps/protocol/src/app/components/SelectBoost.tsx
@@ -3,11 +3,12 @@ import styled from 'styled-components'
 
 import { BoostDirector__factory } from '@apps/artifacts/typechain'
 import { useAccount, useSigner } from '@apps/base/context/account'
-import { useDataState, useSelectedMassetState, BoostedSavingsVaultState } from '@apps/base/context/data'
+import { useDataState, BoostedSavingsVaultState } from '@apps/base/context/data'
 import { usePropose } from '@apps/base/context/transactions'
 import { ViewportWidth } from '@apps/base/theme'
 import { TransactionManifest, Interfaces } from '@apps/transaction-manifest'
 import { Button } from '@apps/components/core'
+import { useSelectedMassetState } from '@apps/hooks'
 
 interface Props {
   vault?: BoostedSavingsVaultState
