+++ apps/protocol/src/app/context/SelectedSaveVersionProvider.tsx
@@ -1,9 +1,10 @@
 import React, { createContext, Dispatch, FC, SetStateAction, useContext, useEffect, useRef, useState } from 'react'
 
-import { useSelectedMassetState, SavingsContractState } from '@apps/base/context/data'
+import { SavingsContractState } from '@apps/base/context/data'
 import { useSelectedMassetName } from '@apps/base/context/masset'
 import { useWalletAddress } from '@apps/base/context/account'
 import { useV1SavingsBalanceQuery } from '@apps/artifacts/graphql/protocol'
+import { useSelectedMassetState } from '@apps/hooks'
 
 export enum SaveVersion {
   V1 = 1,
