+++ apps/protocol/src/app/hooks/useMassetExchangeRate.ts
@@ -1,6 +1,6 @@
+import { useSelectedMassetState } from '@apps/hooks'
 import { useMemo } from 'react'
 
-import { useSelectedMassetState } from '@apps/base/context/data'
 import { BigDecimal } from '@apps/bigdecimal'
 import type { MassetState } from '@apps/base/context/data'
 import type { FetchState, BigDecimalInputValue, BigDecimalInputValues } from '@apps/hooks'
