+++ apps/protocol/src/app/hooks/useSelectedMassetPrice.ts
@@ -1,6 +1,5 @@
-import { useSelectedMassetState } from '@apps/base/context/data'
 import { useFetchPriceCtx } from '@apps/base/context/prices'
-import { FetchState } from '@apps/hooks'
+import { FetchState, useSelectedMassetState } from '@apps/hooks'
 
 export const useSelectedMassetPrice = (): FetchState<number> => {
   const massetState = useSelectedMassetState()
