+++ apps/protocol/src/app/pages/Exchange/Mint/MintMasset.tsx
@@ -5,11 +5,11 @@ import { FeederPool__factory, Masset__factory } from '@apps/artifacts/typechain'
 
 import { usePropose } from '@apps/base/context/transactions'
 import { useSigner, useWalletAddress } from '@apps/base/context/account'
-import { useSelectedMassetState, MassetState } from '@apps/base/context/data'
+import { MassetState } from '@apps/base/context/data'
 import { useTokenSubscription } from '@apps/base/context/tokens'
 import { useChainIdCtx } from '@apps/base/context/network'
 
-import { useBigDecimalInput, useSlippage, BigDecimalInputValue, useMinimumOutput } from '@apps/hooks'
+import { useBigDecimalInput, useSlippage, BigDecimalInputValue, useMinimumOutput, useSelectedMassetState } from '@apps/hooks'
 import { TransactionManifest, Interfaces } from '@apps/transaction-manifest'
 import { AssetInput, SendButton } from '@apps/components/forms'
 import { Arrow, ExchangeRate, TransactionInfo } from '@apps/components/core'
