+++ apps/protocol/src/app/pages/Exchange/Redeem/RedeemExactBassets.tsx
@@ -3,12 +3,12 @@ import React, { FC, useMemo } from 'react'
 import { Masset__factory } from '@apps/artifacts/typechain'
 import { usePropose } from '@apps/base/context/transactions'
 import { useSigner, useWalletAddress } from '@apps/base/context/account'
-import { useSelectedMassetState, MassetState } from '@apps/base/context/data'
+import { MassetState } from '@apps/base/context/data'
 import { useTokens, useTokenSubscription } from '@apps/base/context/tokens'
 import { BigDecimal } from '@apps/bigdecimal'
 import { TransactionManifest, Interfaces } from '@apps/transaction-manifest'
 import { SendButton, MultiAssetExchangeProvider, OneToManyAssetExchange, useMultiAssetExchangeState } from '@apps/components/forms'
-import { useMaximumOutput } from '@apps/hooks'
+import { useMaximumOutput, useSelectedMassetState } from '@apps/hooks'
 
 import { useSelectedMassetPrice } from '../../../hooks/useSelectedMassetPrice'
 import { Route, useEstimatedOutputMulti } from '../../../hooks/useEstimatedOutputMulti'
