+++ apps/protocol/src/app/pages/Exchange/Redeem/RedeemMasset.tsx
@@ -4,10 +4,10 @@ import styled from 'styled-components'
 import { FeederPool__factory, Masset__factory } from '@apps/artifacts/typechain'
 import { usePropose } from '@apps/base/context/transactions'
 import { useSigner, useWalletAddress } from '@apps/base/context/account'
-import { useSelectedMassetState, MassetState } from '@apps/base/context/data'
+import { MassetState } from '@apps/base/context/data'
 import { useTokenSubscription } from '@apps/base/context/tokens'
 
-import { useBigDecimalInput, useSlippage, useMinimumOutput, BigDecimalInputValue } from '@apps/hooks'
+import { useBigDecimalInput, useSlippage, useMinimumOutput, BigDecimalInputValue, useSelectedMassetState } from '@apps/hooks'
 import { TransactionManifest, Interfaces } from '@apps/transaction-manifest'
 
 import { AssetInput, SendButton } from '@apps/components/forms'
