+++ apps/protocol/src/app/pages/Exchange/Swap/index.tsx
@@ -3,10 +3,9 @@ import styled from 'styled-components'
 
 import { FeederPool__factory, Masset__factory } from '@apps/artifacts/typechain'
 import { useSigner, useWalletAddress } from '@apps/base/context/account'
-import { useSelectedMassetState } from '@apps/base/context/data'
 import { useTokenSubscription } from '@apps/base/context/tokens'
 import { usePropose } from '@apps/base/context/transactions'
-import { useBigDecimalInput, useMinimumOutput, BigDecimalInputValue, useSlippage } from '@apps/hooks'
+import { useBigDecimalInput, useMinimumOutput, BigDecimalInputValue, useSlippage, useSelectedMassetState } from '@apps/hooks'
 import { AddressOption } from '@apps/types'
 import { TransactionManifest, Interfaces } from '@apps/transaction-manifest'
 import { AssetSwap, SendButton } from '@apps/components/forms'
