+++ apps/protocol/src/app/pages/Pools/Detail/UserRewards.tsx
@@ -7,10 +7,10 @@ import { useToggle } from 'react-use'
 import { BoostedSavingsVault__factory, BoostedSavingsVault } from '@apps/artifacts/typechain'
 import { usePropose } from '@apps/base/context/transactions'
 import { useIsMasquerading, useSigner } from '@apps/base/context/account'
-import { useSelectedMassetState } from '@apps/base/context/data'
 import { TransactionManifest, Interfaces } from '@apps/transaction-manifest'
 import { SendButton } from '@apps/components/forms'
 import { CountUp, Table, TableCell, TableRow, Button } from '@apps/components/core'
+import { useSelectedMassetState } from '@apps/hooks'
 
 import { useSelectedSaveVersion } from '../../../context/SelectedSaveVersionProvider'
 import { StreamType, useRewardStreams } from '../../../context/RewardStreamsProvider'
