+++ apps/protocol/src/app/pages/Pools/cards/AssetCard.tsx
@@ -7,13 +7,14 @@ import Skeleton from 'react-loading-skeleton'
 import { useFeederPoolMetricsQuery } from '@apps/artifacts/graphql/feeders'
 import { useSelectedMassetName } from '@apps/base/context/masset'
 import { useTokenSubscription } from '@apps/base/context/tokens'
-import { FeederPoolState, useFeederPool } from '@apps/base/context/data'
+import { FeederPoolState } from '@apps/base/context/data'
 import { useBlockNumbers } from '@apps/base/context/block'
 import { BigDecimal } from '@apps/bigdecimal'
 import { ViewportWidth } from '@apps/base/theme'
 import { toK } from '@apps/formatters'
 import { CountUp, CountUpUSD, Tooltip } from '@apps/components/core'
 import { TokenIcon, TokenPair } from '@apps/components/icons'
+import { useFeederPool } from '@apps/hooks'
 
 import { useSelectedMassetPrice } from '../../../hooks/useSelectedMassetPrice'
 import { useFeederPoolApy } from '../../../hooks/useFeederPoolApy'
