+++ apps/protocol/src/app/pages/Save/v1/SaveMigrationProvider.tsx
@@ -4,12 +4,12 @@ import React, { createContext, FC, useCallback, useContext, useMemo, useState } 
 import { BigNumber, constants } from 'ethers'
 
 import { ERC20__factory, ISavingsContractV1, ISavingsContractV1__factory } from '@apps/artifacts/typechain'
-import { useSelectedMassetState } from '@apps/base/context/data'
 import { useTokenAllowance } from '@apps/base/context/tokens'
 import { useSigner } from '@apps/base/context/account'
 import { Transaction, useTransactionsDispatch, useTransactionsState } from '@apps/base/context/transactions'
 import { TransactionManifest, TransactionStatus, Interfaces } from '@apps/transaction-manifest'
 import { StepProps } from '@apps/components/core'
+import { useSelectedMassetState } from '@apps/hooks'
 
 import { useSelectedSaveVersion } from '../../../context/SelectedSaveVersionProvider'
