+++ apps/protocol/src/app/pages/Save/v2/SaveDeposit.tsx
@@ -4,11 +4,11 @@ import { BoostedSavingsVault__factory, ISavingsContractV2__factory, SaveWrapper_
 
 import { useSigner } from '@apps/base/context/account'
 import { usePropose } from '@apps/base/context/transactions'
-import { useSelectedMassetState, MassetState } from '@apps/base/context/data'
+import { MassetState } from '@apps/base/context/data'
 import { useNetworkAddresses, useNetworkPrices } from '@apps/base/context/network'
 import { useNativeToken, useTokenSubscription } from '@apps/base/context/tokens'
 
-import { useSlippage, useBigDecimalInput } from '@apps/hooks'
+import { useSlippage, useBigDecimalInput, useSelectedMassetState } from '@apps/hooks'
 import { BigDecimal } from '@apps/bigdecimal'
 import { TransactionManifest, Interfaces } from '@apps/transaction-manifest'
 import { getPenaltyPercentage } from '@apps/quick-maths'
