+++ apps/protocol/src/app/pages/Save/v2/SavePosition.tsx
@@ -1,10 +1,11 @@
 import React, { FC, useMemo } from 'react'
 import styled from 'styled-components'
 
-import { useSelectedMassetState, MassetState } from '@apps/base/context/data'
+import { MassetState } from '@apps/base/context/data'
 import { ViewportWidth } from '@apps/base/theme'
 import { BigDecimal } from '@apps/bigdecimal'
 import { CountUp } from '@apps/components/core'
+import { useSelectedMassetState } from '@apps/hooks'
 
 import { useSelectedMassetPrice } from '../../../hooks/useSelectedMassetPrice'
 import { useSelectedSaveVersion } from '../../../context/SelectedSaveVersionProvider'
