+++ apps/protocol/src/app/pages/Save/v2/useSaveOutput.ts
@@ -3,11 +3,11 @@ import { Signer } from 'ethers'
 import { useEffect } from 'react'
 import { useDebounce } from 'react-use'
 import { IUniswapV2Router02__factory, FeederPool__factory, Masset__factory, SaveWrapper__factory } from '@apps/artifacts/typechain'
-import { useSelectedMassetState, MassetState } from '@apps/base/context/data'
+import { MassetState } from '@apps/base/context/data'
 import { useSigner } from '@apps/base/context/account'
 import { AllNetworks, useNetworkAddresses, useNetworkPrices } from '@apps/base/context/network'
 import { useSelectedMassetConfig } from '@apps/base/context/masset'
-import { useFetchState, FetchState } from '@apps/hooks'
+import { useFetchState, FetchState, useSelectedMassetState } from '@apps/hooks'
 import { sanitizeMassetError } from '@apps/formatters'
 import { BigDecimal } from '@apps/bigdecimal'
 import { getPriceImpact } from '@apps/quick-maths'
