+++ apps/protocol/src/app/pages/Stats/index.tsx
@@ -2,9 +2,9 @@ import React, { FC, useLayoutEffect } from 'react'
 import styled from 'styled-components'
 import Skeleton from 'react-loading-skeleton'
 
-import { useSelectedMassetState } from '@apps/base/context/data'
 import { H2, H3, CountUp, ThemedSkeleton } from '@apps/components/core'
 import { ReactComponent as StatsIcon } from '@apps/components/icons/circle/stats.svg'
+import { useSelectedMassetState } from '@apps/hooks'
 
 import { VolumeChart } from '../../components/stats/VolumeChart'
 import { AggregateChart } from '../../components/stats/AggregateChart'
