+++ libs/hooks/src/index.ts
@@ -11,6 +11,8 @@ export * from './lib/useEmojisplosion'
 export * from './lib/useFetchState'
 export * from './lib/useModalComponent'
 export * from './lib/useOutput'
-export * from './lib/usePrice'
 export * from './lib/useSimpleInput'
 export * from './lib/utils'
+export * from './lib/useSelectedMassetState'
+export * from './lib/useSelectedBoostedSavingsVault'
+export * from './lib/useFeederPool'
